moved index into class, sorting index items with accents
