+++ lib/kitchen/patches/nokogiri.rb
